+++ angular-workspace/projects/ni/nimble-angular/table-column/date-text/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-table-column-date-text.directive';
+export * from './nimble-table-column-date-text.module';