+++ packages/nimble-components/src/anchor-tab/index.ts
@@ -46,7 +46,10 @@ export class AnchorTab extends AnchorBase {
 const nimbleAnchorTab = AnchorTab.compose({
     baseName: 'anchor-tab',
     template,
-    styles
+    styles,
+    shadowOptions: {
+        delegatesFocus: true
+    }
 });
 
 DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleAnchorTab());