+++ packages/nimble-components/src/table-column/anchor/tests/table-column-anchor.stories.ts
@@ -19,7 +19,6 @@ import { AnchorAppearance } from '../../../anchor/types';
 const metadata: Meta<SharedTableArgs> = {
     title: 'Incubating/Table Column - Anchor',
     decorators: [withActions],
-    tags: ['autodocs'],
     parameters: {
         actions: {
             handles: sharedTableActions
@@ -75,16 +74,7 @@ interface AnchorColumnTableArgs extends SharedTableArgs {
     underlineHidden: boolean;
 }
 
-const anchorColumnDescription = `The \`nimble-table-column-anchor\` column is used to display string fields as links or text in the \`nimble-table\`. If a row provides an href for a link, that cell will display a link, otherwise it will display plain text.
-
-In an Angular app, you can configure a callback to intercept clicks so that you may invoke the router to perform the navigation instead of the default handler:
-\`\`\`
-<nimble-table-column-anchor [navigationGuard]="doRouterNavigation">
-    Link
-</nimble-table-column-anchor>
-\`\`\`
-The function bound to \`navigationGuard\` should be of type \`NavigationGuard\`. It receives the clicked row id and should return a boolean indicating whether the default click handler should be run. I.e. return \`false\` if you have handled the navigation. The function is only called for left clicks with no modifier keys pressed.
-`;
+const anchorColumnDescription = 'The `nimble-table-column-anchor` column is used to display string fields as links or text in the `nimble-table`. If a row provides an href for a link, that cell will display a link, otherwise it will display plain text.';
 
 export const anchorColumn: StoryObj<AnchorColumnTableArgs> = {
     parameters: {