+++ packages/nimble-components/src/table-column/number-text/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Formatting scheme for the number-text table column
+ */
+export const NumberTextFormat = {
+    default: undefined,
+    roundToInteger: 'round-to-integer'
+} as const;
+export type NumberTextFormat =
+    (typeof NumberTextFormat)[keyof typeof NumberTextFormat];